domain_conf: Read and Write quorum config
Add the capabiltty to libvirt to parse and format the quorum syntax
as described here:
http://www.redhat.com/archives/libvir-list/2014-May/msg00533.html

We need a different unique index for each backing store
to manipulate individually each child of a quorum.

If a disk have 2 childs A and B, the index will be 1 for A and 2 for B,
but if A have 2 childs C and D, so A will have the index 1, C and D: 2 and 3
and B will be 4.

Here is a representation of our disk tree and they indexs:

         C[2]
        /
    A[1]
   /    \D[3]
HDA
   \
    B[4]

Signed-off-by: Matthias Gatto <matthias.gatto@outscale.com>