virstoragefile: Add virStorageSourceIsContener
Add a new function which return true if a virStorageSourcePtr is
a "contener".
A contener is a source, which has more than one source as sources.
example: a quorum on qemu, or any kind of RAID.

For now, quorum is the only "contener" we have.

This function is usefull, because, a lot of code access directly
to a virStorageSource internal member (like path) with some functions
like "virDomainDiskGetSource".
This beavious don't work with quorum, and so we need to add
exeption for these functions, but I'm not convinced by the idea to add a lot
of (disk->format == QUORUM) in all the code that deserve
exeption for QUORUM, so I've add a generic function for this.

Signed-off-by: Matthias Gatto <matthias.gatto@outscale.com>