qemu_driver: qemuDiskPathToAlias handle contener
qemuDiskPathToAlias return 0 when trying to get a alias from
a quorum.
This is due to the fact whan a quorum path is NULL, but still
a quorum is a valide disk.
Moreover, qemuDiskPathToAlias is use in a lot of functions
to get a disk idx, most of these function does not work with
quorum, but this should be explicitely explain, and the error
should not be the one return by qemuDiskPathToAlias.

Signed-off-by: Matthias Gatto <matthias.gatto@outscale.com>