quorum: Block snapshot operation with "contener"
For now we block all snapshot operations with quorum, because it would require
a lot more code, espacially because Qemu doesn't really suport it.

I guess, we can use node-name, and manually snapshoot all qcow from a
virStorageSource and use this as a quorum's snapshot,
but libvirt doesn't support node-name, and we don't need nodename
anymore to use a quorum in qemu.

I actually have some patchs which could partially support quorum snapshot
on my computer, but nothing suitable to be upstream, so I prefer to have
a stable versions of quorum inside libvirt before dealling with
snapshot.

Signed-off-by: Matthias Gatto <matthias.gatto@outscale.com>